refactor: SensorTrendPointResponse double를 BigDecimal로 변경

+++ src/main/java/site/yesaido/cultivation_server/sensor/dto/response/influx/SensorTrendPointResponse.java
@@ -1,9 +1,10 @@
 package site.yesaido.cultivation_server.sensor.dto.response.influx;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 public record SensorTrendPointResponse(
         Instant measuredAt,
-        Double value
+        BigDecimal value
 ) {
 }
